Re-enable server cache
The hope is that having only one server (HTTP, instead of HTTP + HTTPS) would be
enough to relieve the server of the stress it experiences memory-wise.

While there was a crash a week with two servers, there was only one crash in a
month since the switch, and the server was immediately restarted.