[PATCH v3] Add support for issuestats.com
This commit adds support for issuestats.com, as requested in #351.  The
implementation is straight-forward, with the exception of the handling
of 'not found' which currently returns text/html with code 200 (see
hstove/issue_stats#38).  The URL parsing is somewhat lax to allow future
support for other hosting platforms which may not follow GitHub's
user/repo URL scheme.  It also passes through the `concise` query
parameter, if present, which shortens the text a bit.

Changes in v3:
- Fix `concise` option assignment holdover missed in v2 changes.
- Make `label` check null/undefined rather than falsey.

Changes in v2:
- Move type to first parameter and shorten to 1 character.
- Change `concise` option from query param to optional path component
- Make `color` check null/undefined rather than falsey.
- Add try-catch around JSON value extraction and response, by request.

Fixes: #351

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>