Allow color to be expressed as a percentage.
What
===
Allow color to be expressed as a percentage as any 1-3 digit number
followed by an optional decimal point and optional decimals, finally
followed by a percentage symbol. An optional query parameter
`percentageColorType` defines how the percentage is mapped to colors
that supports a default range of 0, 50, 75, 100 and the existing
coverage range of 0, 80, 90, 100.

Why
===
It's useful to abstract the actual colors away from other services that
generate the badge URLs, and instead allow the simple passing of a
percentage.