Fix png image size
On my desktop, using `-density` option of `convert` does change the
resolution and print size of the image, keeping the image geometry.
On Scalingo containers, using `-density` option of `convert` does change the
resolution and geometry, keeping the print size. This results to an
image which is 1.25x bigger than the original svg

The behavior on Scalingo is not the one expected. This behavior is not
present when removing `-density` parameter. I could not find why it
behaves differently on my desktop and scalingo.