Improve search performance
This change applies a unique ID to each badge example, and uses that to key the DOM.

Before this change, the key was the array index. While constant with the filter, it mutated as the filter changed, causing React to mutate the DOM very inefficiently. Badges were filtered out by mutating al the rows instead of simply removing the affected rows.

I first tried a version that applied a class with `display: none` to the hidden elements. That worked; however the implementation has a bunch of indirection, and this works nearly as well.