Rewrite two of the svg badges
This tackles two of the services using `fetchFromSvg`. The others are in codacy, which are broken right now; see #1985.

There’s a design decision to make here about how to structure the shared code. One is with a subclass, and the other with a function.

After working on a few service families, I noticed a pattern. First I could code the first service. Then I would start to code the second, and  begin to separate out the shared code into an abstract superclass.

This involved a lot of typing:

- `require` statements would be moved into the module containing the base class
- new `require` statements would be added to the module with the subclass
- The `extends` line would change
- All the shared code would move

The current arrangement also produced base service code which couldn’t be tested without creating a subclass, which has meant that tests of `fetch` end up being written as service tests, though probably they would be more clearly expressed as unit tests with intercepted http calls.

I also dislike deep inheritance hierarchies because they make it harder for new developers to understand what is happening. They are also difficult to change, on account of all the implicit dependencies.

It would be inconceivable to create a base class for a single service, based on the possibility of writing more services someday. This means going from n=1 to 2 is services per family is (relatively) hard, while going from n >=2 to n+1 is easy.

This led me to want to experiment with composition; hence taking the “function” approach here.