Add pdk badge for puppet forge
This adds a badge that can be used to show whether or not a module was developing using the puppet development kit (pdk) and if so what version it's currently using.

Before this gets merged i would love to get some feedback first from some key puppet employees in order to make adjustments if needed.

/cc @davids @scotje

Closes #2255