BaseService terminology: Rename `url` to `route`
The term “url” is overloaded in services, to refer to the Shields route and also the API URL. Calling the Shields URL a “route” makes it clearer and more obvious which one of these we’re talking about. It’s a small thing, though seems like an improvement.

We have a few functions called `buildUrl`. I’ve renamed them to `buildRoute` when they refer to routes, and left them as `buildUrl` when they refer to API URLs.