change php-v to look for latest stable version
This does make functional changes.

Previously if this badge was called with no version
argument we were assuming 'dev-master'
(an unstable branch which may not exist)

Now we're assuming the latest stable version if one
exists or the latest unstable version otherwise.

This brings us into line with the license badge.