Rewrite test of analytics endpoint
This test is being weirdly flaky in #2809. The problem seems to be in
the test helper code, so I rewrote this using Joi.

I imagine the change has to do with a change to the test ordering. It's
a bit puzzling, though this test seems fine and the endpoint is rarely
used.